Snap Drop Grabbables + Partial fix door jitters
implemented dropping grabbables when our hand is too far from the model of the grabbable, such as the sink handle.

Fixed door jitters, but the implementation for snap dropping  grabbables doesn't work because we grab an invisible object and carry it with us so our distance from the grabbed item is always 0.